backend/requirements_develop.txt
coverage==6.4.1
django-stubs==1.12.0
grpc-stubs==1.24.10
mypy==0.961
pylint==2.14.3
pylint-django==2.5.3
pylint-protobuf==0.20.2
types-protobuf==3.19.22
